Invoices module refactored
