+++ apps/frontend/src/app/app.component.html
@@ -57,7 +57,7 @@
         <div matListItemTitle>Табло за управление</div>
       </mat-list-item>
 
-      <mat-list-item matRipple routerLink="/invoicesList">
+      <mat-list-item matRipple routerLink="/invoices">
         <mat-icon matListItemIcon> wysiwyg </mat-icon>
         <div matListItemTitle>Фактури</div>
       </mat-list-item>
