+++ apps/frontend/src/app/modules/invoices/invoice/invoice.component.scss
@@ -1,0 +1,435 @@
+form {
+  max-width: 90%;
+  margin: 0 auto;
+  padding: 10px;
+}
+
+mat-form-field {
+  width: 100%;
+}
+
+mat-form-field,
+input {
+  font-size: 14px;
+}
+
+button {
+  border: none;
+  cursor: pointer;
+}
+.hidden {
+  display: none !important;
+}
+button:hover {
+  opacity: 0.8;
+}
+table {
+  width: 99% !important;
+  margin-top: 10px;
+  margin-bottom: 10px;
+  margin-left: -15px !important;
+}
+
+tr td {
+  padding-right: 10px !important;
+}
+
+mat-icon {
+  color: rgb(48, 48, 48);
+}
+mat-card {
+  margin-top: 20px;
+}
+.mat-card-header-text {
+  margin: 0px !important;
+}
+
+.container {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-template-rows: 1fr;
+  gap: 0px 1px;
+  grid-template-areas: '. .';
+  column-gap: 10px;
+}
+
+.col-1,
+.col-2 {
+  margin-top: 10px;
+  margin-bottom: 0px;
+  padding-top: 30px !important;
+}
+.radioBtn {
+  margin: 10px !important;
+}
+
+.btn-submit {
+  margin-top: 10px;
+  margin-left: 10px;
+}
+.mat-elevation-z8 {
+  box-shadow: none !important;
+}
+
+$table-header: #fff;
+$table-header-border: #fff;
+$table-border: #ccc;
+
+div {
+  box-sizing: border-box;
+}
+.table-container {
+  display: block;
+  margin: 2em auto;
+}
+.flex-table {
+  display: flex;
+  flex-flow: row wrap;
+
+  &:first-of-type .flex-row {
+    font-weight: 600;
+    color: rgb(48, 48, 48);
+  }
+}
+.flex-row {
+  width: calc(100% / 6);
+  text-align: right;
+  padding: 0.5em 0.5em;
+  border-bottom: solid 1px $table-border;
+  color: rgb(104, 102, 102);
+  height: 57px !important;
+}
+
+.column {
+  display: flex;
+  flex-flow: column wrap;
+  width: 95%;
+  padding: 0;
+  .flex-row {
+    display: flex;
+    flex-flow: row wrap;
+    width: 100%;
+    padding: 0;
+    border: 0;
+    border-bottom: solid 1px $table-border;
+    color: #ccc;
+  }
+}
+.flex-cell {
+  width: calc(100% / 3);
+  text-align: center;
+  padding: 0.5em 0.5em;
+  &:last-child {
+    border-right: 0px;
+  }
+}
+
+.table-container .header .flex-row {
+  text-align: left;
+}
+.table-container .right {
+  text-align: right !important;
+  height: 57px !important;
+  padding-top: 25px;
+}
+.btn-submit {
+  background-color: #172647;
+  color: #f4f2f1;
+}
+.icon-add-btn {
+  color: rgb(48, 48, 48);
+  margin-top: -7px;
+  margin-left: 7px;
+}
+.flex-row.first {
+  width: 7%;
+}
+.flex-row.second,
+.nameFieldHeader {
+  width: 45%;
+}
+.flex-row.third,
+.qtyHeader {
+  width: 10%;
+}
+.flex-row.fourth,
+.measureHeader {
+  width: 13%;
+}
+.flex-row.fifth,
+.unitPriceHeader {
+  width: 13%;
+}
+.flex-row.sixth,
+.amountHeader {
+  width: 12%;
+}
+.amountHeader {
+  text-align: start !important;
+}
+.amount,
+.vatPercent {
+  color: rgb(104, 102, 102);
+}
+.flex-row .nameFieldHeader {
+  width: 30%;
+}
+.vatReason {
+  width: 100%;
+}
+.wayOfPaying {
+  margin-top: 25px;
+  width: 18.5%;
+  width: 28.8%;
+  text-align: start;
+}
+.amount {
+  text-align: end;
+  width: 26.3%;
+}
+.vatPercent,
+.price,
+.sum {
+  width: 79%;
+}
+.vat,
+.priceNum {
+  width: 21%;
+}
+.header {
+  height: 57px !important;
+  margin: 0;
+  padding: 0;
+}
+.amountHeader,
+.nameFieldHeader,
+.qtyHeader,
+.measureHeader,
+.unitPriceHeader {
+  padding-top: 20px;
+}
+.icon-add-btn {
+  padding-top: 8px;
+}
+.table-container {
+  margin-top: -10px !important;
+}
+.measure-select {
+  margin-top: -2px;
+}
+
+.title {
+  margin-bottom: 20px;
+  font-weight: 600;
+  color: rgb(48, 48, 48);
+}
+.grid-layout {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  grid-template-rows: repeat(3, 1fr);
+  grid-column-gap: 0px;
+  grid-row-gap: 10px;
+  margin-left: 10px;
+  column-gap: 30px;
+}
+.div1 {
+  grid-area: 1 / 1 / 2 / 2;
+}
+.div2 {
+  grid-area: 1 / 2 / 2 / 3;
+}
+.div3 {
+  grid-area: 1 / 3 / 2 / 4;
+}
+.div4 {
+  grid-area: 1 / 4 / 2 / 5;
+}
+.div5 {
+  grid-area: 2 / 1 / 3 / 2;
+}
+.div6 {
+  grid-area: 2 / 2 / 3 / 3;
+}
+.div7 {
+  grid-area: 2 / 3 / 3 / 4;
+}
+.grid-layout-top {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  grid-column: span 2;
+  margin-bottom: 20px;
+}
+
+::ng-deep .mat-form-field-appearance-outline .mat-form-field-infix {
+  padding: 0.9em 0 0.8em 0 !important;
+  margin-top: -4px;
+}
+@media all and (max-width: 980px) {
+  .flex-row.second,
+  .nameFieldHeader {
+    width: 40%;
+  }
+  .flex-row.third,
+  .qtyHeader {
+    width: 15%;
+  }
+  .currency {
+    margin-left: 0px !important;
+  }
+  .datepicker {
+    margin-left: 0px !important;
+  }
+  .basis {
+    height: 100px !important;
+  }
+}
+@media all and (max-width: 767px) {
+  .div1 {
+    grid-area: 1 / 1 / 2 / 2;
+  }
+  .div2 {
+    grid-area: 1 / 2 / 2 / 3;
+  }
+  .div3 {
+    grid-area: 1 / 3 / 2 / 4;
+  }
+  .div4 {
+    grid-area: 2 / 1 / 3 / 2;
+  }
+  .div5 {
+    grid-area: 2 / 2 / 3 / 3;
+  }
+  .div6 {
+    grid-area: 3 / 1 / 4 / 2;
+  }
+  .div7 {
+    grid-area: 3 / 2 / 4 / 3;
+  }
+  .amountHeader,
+  .nameFieldHeader,
+  .qtyHeader,
+  .measureHeader,
+  .unitPriceHeader {
+    display: none;
+  }
+  .amount {
+    text-align: right;
+    border-bottom: 2px solid rgb(48, 48, 48) !important;
+  }
+  .flex-table {
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: center;
+    border-bottom: 0px solid;
+  }
+  .flex-row {
+    width: calc(100% / 2.7);
+    border-bottom: 2px solid rgb(48, 48, 48);
+
+    &.first {
+      width: 100%;
+    }
+  }
+
+  .column {
+    width: 100%;
+  }
+}
+
+@media all and (max-width: 530px) {
+  .currency {
+    margin-left: 0px !important;
+  }
+  .datepicker {
+    margin-left: 0px !important;
+  }
+  .div1 {
+    grid-area: 1 / 1 / 2 / 2;
+  }
+  .div2 {
+    grid-area: 1 / 2 / 2 / 3;
+  }
+  .div3 {
+    grid-area: 2 / 1 / 3 / 2;
+  }
+  .div4 {
+    grid-area: 2 / 2 / 3 / 3;
+  }
+  .div5 {
+    grid-area: 3 / 1 / 4 / 2;
+  }
+  .div6 {
+    grid-area: 3 / 2 / 4 / 3;
+  }
+  .div7 {
+    grid-area: 4 / 1 / 5 / 2;
+  }
+  .amount {
+    border-bottom: 2px solid rgb(48, 48, 48) !important;
+  }
+  .amountHeader,
+  .nameFieldHeader,
+  .qtyHeader,
+  .measureHeader,
+  .unitPriceHeader {
+    display: none;
+  }
+  .flex-table {
+    .flex-row {
+      border-bottom: 0;
+      width: 100%;
+    }
+    .flex-row:last-of-type {
+      border-bottom: solid 1px $table-border;
+    }
+  }
+
+  .header {
+    .flex-row {
+      border-bottom: solid 1px;
+      width: 100% !important;
+    }
+  }
+
+  .flex-row {
+    width: 100%;
+
+    &.first {
+      width: 100%;
+      border-bottom: solid 1px $table-border;
+    }
+  }
+
+  .column {
+    width: 100%;
+    .flex-row {
+      border-bottom: solid 1px $table-border;
+    }
+  }
+
+  .flex-cell {
+    width: 100%;
+  }
+}
+@media (max-width: 600px) {
+  .container {
+    grid-template-columns: 1fr;
+    grid-template-areas:
+      'col-1'
+      'col-2';
+  }
+  .grid-layout-top {
+    grid-template-columns: repeat(1, 1fr);
+    grid-template-rows: repeat(4, 1fr);
+    grid-column: span 1;
+  }
+}
+.btn-submit {
+  background-color: #172647;
+  color: #f4f2f1;
+  margin-top: 10px;
+}
+.error {
+  color: rgb(223, 59, 59);
+  font-weight: 500;
+}
