--- apps/frontend/src/app/modules/invoicesList/invoicesList.component.html
@@ -1,68 +1,0 @@
-<div class="mat-elevation-z8">
-  <h1>Фактури</h1>
-  <button
-    class="btn-add mat-raised-button matTooltip"
-    aria-label="добавяне на контрагент"
-    mat-raised-button
-  >
-    <a routerLink="/invoices">
-      <span class="add-text">Добавяне</span>
-    </a>
-  </button>
-
-  <table mat-table [dataSource]="dataSource" class="">
-    <!-- Name Column -->
-    <ng-container matColumnDef="id">
-      <th mat-header-cell *matHeaderCellDef>ID</th>
-      <td mat-cell *matCellDef="let invoice">{{ invoice.id }}</td>
-    </ng-container>
-
-    <!-- City Column -->
-    <ng-container matColumnDef="name">
-      <th mat-header-cell *matHeaderCellDef>Фирма</th>
-      <td mat-cell *matCellDef="let invoice">{{ invoice.receiver }}</td>
-    </ng-container>
-
-    <!-- MOL Column -->
-    <ng-container matColumnDef="date">
-      <th mat-header-cell *matHeaderCellDef>Дата</th>
-      <td mat-cell *matCellDef="let invoice">{{ invoice.issue_date }}</td>
-    </ng-container>
-
-    <!-- EGN Column -->
-    <ng-container matColumnDef="amount">
-      <th mat-header-cell *matHeaderCellDef>Стойност</th>
-      <td mat-cell *matCellDef="let invoice">
-        <div *ngFor="let item of invoice.items">
-          {{ item.price }} {{ invoice.currency.currencyCode }}
-        </div>
-      </td>
-    </ng-container>
-
-    <!-- Tools Column -->
-    <ng-container matColumnDef="tools">
-      <th mat-header-cell *matHeaderCellDef>Инструменти</th>
-      <td mat-cell *matCellDef="let invoice">
-        <button
-          class="icon-edit-btn"
-          mat-icon-button
-          matTooltip="редактиране на ред"
-          aria-label="редактиране на ред"
-          (click)="editInvoice(invoice)"
-        >
-          <mat-icon>edit</mat-icon>
-        </button>
-      </td>
-    </ng-container>
-
-    <!-- Table rows -->
-    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-    <tr mat-row *matRowDef="let contractor; columns: displayedColumns"></tr>
-  </table>
-  <mat-paginator
-    [length]="invoices.length"
-    [pageSize]="5"
-    [pageSizeOptions]="[5, 10, 25, 100]"
-    showFirstLastButtons
-  ></mat-paginator>
-</div>
