--- apps/frontend/src/app/modules/invoicesList/invoicesList.component.scss
@@ -1,25 +1,0 @@
-table {
-  width: 100%;
-}
-.btn-add {
-  background-color: #172647;
-
-  float: right;
-  margin-bottom: 10px;
-}
-.btn-add a {
-  color: #f4f2f1;
-  text-decoration: none;
-  cursor: pointer;
-}
-h1 {
-  color: #172647;
-  font-weight: 600;
-}
-.mat-elevation-z8 {
-  background-color: #fff;
-  padding: 20px;
-}
-.btn-add {
-  margin-top: 50px;
-}
