--- apps/frontend/src/app/modules/invoicesList/invoicesList.module.ts
@@ -1,41 +1,0 @@
-import { InvoicesListComponent } from './invoicesList.component';
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { Routes, RouterModule } from '@angular/router';
-import { MatCardModule } from '@angular/material/card';
-import { MatButtonModule } from '@angular/material/button';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatSliderModule } from '@angular/material/slider';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatIconModule } from '@angular/material/icon';
-import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatNativeDateModule } from '@angular/material/core';
-import { MatTableModule } from '@angular/material/table';
-import { MatTooltipModule } from '@angular/material/tooltip';
-import { MatPaginatorModule } from '@angular/material/paginator';
-
-const routes: Routes = [{ path: '', component: InvoicesListComponent }];
-
-@NgModule({
-  declarations: [InvoicesListComponent],
-  imports: [
-    CommonModule,
-    RouterModule.forChild(routes),
-    MatCardModule,
-    MatButtonModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatDatepickerModule,
-    MatSliderModule,
-    MatMenuModule,
-    MatIconModule,
-    MatCheckboxModule,
-    MatNativeDateModule,
-    MatTableModule,
-    MatTooltipModule,
-    MatPaginatorModule,
-  ],
-})
-export class InvoicesListModule {}
